requirements.txt
# PyWebView - 轻量级桌面应用框架
pywebview>=4.0

# CEF (Chromium Embedded Framework) - 提供 Chromium 浏览器内核
# 注意：CEF 是可选依赖，如果安装失败，PyWebView 会使用其他后端
# cefpython3>=66.0  # 如果需要使用 CEF，取消注释并安装

# HTTP 请求
requests>=2.28

# 定时任务调度
apscheduler>=3.10

# 日志
loguru>=0.7

# 数据库
sqlalchemy>=2.0

# 数据验证
pydantic>=2.0

# 加密（用于密码存储）
cryptography>=41.0

# 邮件发送
# smtplib 是 Python 标准库，无需安装

# 数据处理（QMT行情获取需要）
pandas>=2.0

# 行情数据获取
akshare>=1.12.0

# QMT SDK (需要从迅投官方获取并放置到项目目录)
# xtquant - 请参考 MiniQMT 安装说明

# 打包工具（开发时使用）
# pyinstaller>=5.0

